+++ src/main/java/com/project/global/config/PushConfig.java
@@ -1,0 +1,69 @@
+package com.project.global.config;
+
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+import java.security.GeneralSecurityException;
+import java.security.Security;
+
+import org.apache.http.client.config.RequestConfig;
+import org.apache.http.conn.DnsResolver;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.bouncycastle.jce.provider.BouncyCastleProvider;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.project.global.util.NetworkValidator;
+
+import nl.martijndwars.webpush.PushService;
+
+@Configuration
+public class PushConfig {
+
+    @Value("${vapid.key.public}")
+    private String vapidPublicKey;
+
+    @Value("${vapid.key.private}")
+    private String vapidPrivateKey;
+
+    @Value("${push-client.connect-timeout:5000}")
+    private int pushConnectTimeout;
+
+    @Value("${push-client.socket-timeout:5000}")
+    private int pushSocketTimeout;
+
+    @Bean
+    public CloseableHttpClient pushHttpClient() {
+        RequestConfig requestConfig =
+                RequestConfig.custom()
+                        .setConnectTimeout(pushConnectTimeout)
+                        .setSocketTimeout(pushSocketTimeout)
+                        .build();
+
+        DnsResolver ssrfSafeDnsResolver =
+                host -> {
+                    InetAddress[] addresses = InetAddress.getAllByName(host);
+                    for (InetAddress addr : addresses) {
+                        if (NetworkValidator.isInternalAddress(addr)) {
+                            throw new UnknownHostException("Blocked internal address: " + host);
+                        }
+                    }
+                    return addresses;
+                };
+
+        return HttpClientBuilder.create()
+                .setDefaultRequestConfig(requestConfig)
+                .disableRedirectHandling()
+                .setDnsResolver(ssrfSafeDnsResolver)
+                .build();
+    }
+
+    @Bean
+    public PushService pushService() throws GeneralSecurityException {
+        if (Security.getProvider(BouncyCastleProvider.PROVIDER_NAME) == null) {
+            Security.addProvider(new BouncyCastleProvider());
+        }
+        return new PushService(vapidPublicKey, vapidPrivateKey);
+    }
+}
